troca da variavel float pela variavel double

+++ CartasSuperTrunfo.c
@@ -7,24 +7,24 @@ int main() {
   char codigo[6];                   // Armazena o código da carta (como uma string de até 5 caracteres + '\0')
   char cidade[30];                  // Armazena o nome da cidade (até 29 caracteres + '\0')
   unsigned long int populacao = 0;  // Armazena o número de habitantes da cidade
-  float area = 0;                   // Armazena a área da cidade em km²
-  float pib = 0;                    // Armazena o PIB da cidade em bilhões de reais
+  double area = 0;                  // Armazena a área da cidade em km²
+  double pib = 0;                   // Armazena o PIB da cidade em bilhões de reais
   int turistico = 0;                // Armazena a quantidade de pontos turísticos da cidade
-  float densidade = 0;              // Armazena a densidade da população pela areá da cidade
-  float pibPerCapita = 0;           // Armazena o PIB per capita da cidade
-  float superPoder = 0;             // Super Poder da carta
+  double densidade = 0;             // Armazena a densidade da população pela areá da cidade
+  double pibPerCapita = 0;          // Armazena o PIB per capita da cidade
+  double superPoder = 0;            // Super Poder da carta
 
   // Declaração de variáveis da carta 2
   char Estado2;                      // Armazena a letra que representa o estado (de 'A' a 'H')
   char codigo2[6];                   // Armazena o código da carta (como uma string de até 5 caracteres + '\0')
   char cidade2[30];                  // Armazena o nome da cidade (até 29 caracteres + '\0')
   unsigned long int populacao2 = 0;  // Armazena o número de habitantes da cidade 2
-  float area2 = 0;                   // Armazena a área da cidade em km²
-  float pib2 = 0;                    // Armazena o PIB da cidade em bilhões de reais
+  double area2 = 0;                  // Armazena a área da cidade em km²
+  double pib2 = 0;                   // Armazena o PIB da cidade em bilhões de reais
   int turistico2 = 0;                // Armazena a quantidade de pontos turísticos da cidade 2
-  float densidade2 = 0;              // Armazena a densidade da população pela areá da cidade 2
-  float pibPerCapita2 = 0;           // Armazena o PIB per capita da cidade 2
-  float superPoder2 = 0;             // Super Poder da carta 2
+  double densidade2 = 0;             // Armazena a densidade da população pela areá da cidade 2
+  double pibPerCapita2 = 0;          // Armazena o PIB per capita da cidade 2
+  double superPoder2 = 0;            // Super Poder da carta 2
 
 
   // Entrada dos dados da carta 1
@@ -52,11 +52,11 @@ int main() {
 
   // Solicita ao usuário a área da cidade
   printf("Digite a area da cidade em km² da carta 1: ");
-  scanf("%f", &area);  // Lê um número decimal e armazena em area
+  scanf("%lf", &area);  // Lê um número decimal e armazena em area
 
   // Solicita ao usuário o PIB da cidade
   printf("Digite o Produto Interno Bruto da cidade da cidade da carta 1 (em bilhões de reais): ");
-  scanf("%f", &pib);  // Lê um número decimal e armazena em pib
+  scanf("%lf", &pib);  // Lê um número decimal e armazena em pib
 
   // Solicita ao usuário o número de pontos turísticos da cidade
   printf("Digite a quantidade de pontos turísticos na cidade da carta 1: ");
@@ -95,11 +95,11 @@ int main() {
 
   // Solicita ao usuário a área da cidade
   printf("Digite a area da cidade em km² da carta 2: ");
-  scanf("%f", &area2);  // Lê um número decimal e armazena em area2
+  scanf("%lf", &area2);  // Lê um número decimal e armazena em area2
 
   // Solicita ao usuário o PIB da cidade
   printf("Digite o Produto Interno Bruto da cidade da carta 2 (em bilhões de reais): ");
-  scanf("%f", &pib2);  // Lê um número decimal e armazena em pib2
+  scanf("%lf", &pib2);  // Lê um número decimal e armazena em pib2
 
   // Solicita ao usuário o número de pontos turísticos da cidade
   printf("Digite a quantidade de pontos turísticos na cidade da carta 2: ");
@@ -118,30 +118,30 @@ int main() {
 
   // Exibe todos os dados armazenados da carta 1
   printf("\nCarta 1 \n");
-  printf("Estado: %c \n", Estado);                              // Mostra a letra do estado
-  printf("Código: %c%s\n", Estado, codigo);                     // Mostra o código
-  printf("Nome da Cidade: %s\n", cidade);                       // Mostra o nome da cidade
-  printf("População: %lu \n", populacao);                       // Mostra a população
-  printf("Área: %.2f km²\n", area);                             // Mostra a área com duas casas decimais
-  printf("PIB: %.2f bilhões de reais\n", pib);                  // Mostra o PIB com duas casas decimais
-  printf("Número de Pontos Turísticos: %d\n", turistico);       // Mostra o número de pontos turísticos
-  printf("Densidade Populacional: %.2f hab/km²\n", densidade);  // Mostra a densidade populacional da cidade
-  printf("PIB per Capita: R$ %.2f\n", pibPerCapita);            // Mostra o PIB per capita da cidade
-  printf("Super Poder: %.2f\n\n", superPoder);                  // Mostra o Super Poder da cidade
+  printf("Estado: %c \n", Estado);                               // Mostra a letra do estado
+  printf("Código: %c%s\n", Estado, codigo);                      // Mostra o código
+  printf("Nome da Cidade: %s\n", cidade);                        // Mostra o nome da cidade
+  printf("População: %lu \n", populacao);                        // Mostra a população
+  printf("Área: %.2lf km²\n", area);                             // Mostra a área com duas casas decimais
+  printf("PIB: %.2lf bilhões de reais\n", pib);                  // Mostra o PIB com duas casas decimais
+  printf("Número de Pontos Turísticos: %d\n", turistico);        // Mostra o número de pontos turísticos
+  printf("Densidade Populacional: %.2lf hab/km²\n", densidade);  // Mostra a densidade populacional da cidade
+  printf("PIB per Capita: R$ %.2lf\n", pibPerCapita);            // Mostra o PIB per capita da cidade
+  printf("Super Poder: %.2lf\n\n", superPoder);                  // Mostra o Super Poder da cidade
 
   //Exibição dos dados da carta 2
 
   printf("Carta 2\n");
-  printf("Estado: %c\n", Estado2);                               // Mostra a letra do estado2
-  printf("Código: %c%s\n", Estado2, codigo2);                    // Mostra o código2
-  printf("Nome da Cidade: %s\n", cidade2);                       // Mostra o nome da cidade2
-  printf("População: %lu\n", populacao2);                        // Mostra a população2
-  printf("Área: %.2f km²\n", area2);                             // Mostra a área2 com duas casas decimais
-  printf("PIB: %.2f bilhões de reais\n", pib2);                  // Mostra o PIB2 com duas casas decimais
-  printf("Número de Pontos Turísticos: %d\n", turistico2);       // Mostra o número de pontos turísticos
-  printf("Densidade Populacional: %.2f hab/km²\n", densidade2);  // Mostra a densidade populacional da cidade2
-  printf("PIB per Capita: R$ %.2f\n", pibPerCapita2);            // Mostra o PIB per capita da cidade2
-  printf("Super Poder: %.2f\n\n", superPoder2);                  // Mostra o Super Poder da cidade 2
+  printf("Estado: %c\n", Estado2);                                // Mostra a letra do estado2
+  printf("Código: %c%s\n", Estado2, codigo2);                     // Mostra o código2
+  printf("Nome da Cidade: %s\n", cidade2);                        // Mostra o nome da cidade2
+  printf("População: %lu\n", populacao2);                         // Mostra a população2
+  printf("Área: %.2lf km²\n", area2);                             // Mostra a área2 com duas casas decimais
+  printf("PIB: %.2lf bilhões de reais\n", pib2);                  // Mostra o PIB2 com duas casas decimais
+  printf("Número de Pontos Turísticos: %d\n", turistico2);        // Mostra o número de pontos turísticos
+  printf("Densidade Populacional: %.2lf hab/km²\n", densidade2);  // Mostra a densidade populacional da cidade2
+  printf("PIB per Capita: R$ %.2lf\n", pibPerCapita2);            // Mostra o PIB per capita da cidade2
+  printf("Super Poder: %.2lf\n\n", superPoder2);                  // Mostra o Super Poder da cidade 2
 
   char *resultado[] = { "Carta 2 venceu", "Carta 1 venceu" };  // Array com os textos das comparações
